+++ R/ops.R
@@ -17,7 +17,6 @@
 #' all.equal(x * 10, 10 * x)
 #' @export
 Ops.stars <- function(e1, e2) {
-	#ret = if (is.array(e2))
 	ret = if (missing(e2))
 			lapply(e1, .Generic)
 		else if (!inherits(e2, "stars"))
@@ -29,6 +28,8 @@ Ops.stars <- function(e1, e2) {
 			} else
 				mapply(.Generic, e1, e2, SIMPLIFY = FALSE)
 		}
+	if (any(sapply(ret, function(x) is.null(dim(x))))) # happens if e1[[1]] is a factor; #304
+		ret = lapply(ret, function(x) { dim(x) = dim(e1); x })
 	if (! inherits(e1, "stars"))
 		setNames(st_as_stars(ret, dimensions = st_dimensions(e2)), names(e2))
 	else
@@ -59,13 +60,16 @@ Math.stars = function(x, ...) {
 Ops.stars_proxy <- function(e1, e2) {
 	if (!inherits(e1, "stars_proxy"))
 		stop("first argument in expression needs to be the stars_proxy object") # FIXME: needed?? #nocov
-	collect(e1, match.call(), .Generic, "e1")
+	if (missing(e2))
+		collect(e1, match.call(), .Generic, "e1", env = environment())
+	else
+		collect(e1, match.call(), .Generic, c("e1", "e2"), env = environment())
 }
 
 #' @name ops_stars
 #' @export
 Math.stars_proxy = function(x, ...) {
-	collect(x, match.call(), .Generic)
+	collect(x, match.call(), .Generic, env = environment())
 }
 
 
@@ -84,6 +88,7 @@ st_apply = function(X, MARGIN, FUN, ...) UseMethod("st_apply")
 #' @param PROGRESS logical; if \code{TRUE}, use \code{pbapply::pbapply} to show progress bar
 #' @param FUTURE logical;if \code{TRUE}, use \code{future.apply::future_apply} 
 #' @param rename logical; if \code{TRUE} and \code{X} has only one attribute and \code{FUN} is a simple function name, rename the attribute of the returned object to the function name
+#' @param .fname function name for the new attribute name (if one or more dimensions are reduced) or the new dimension (if a new dimension is created); if missing, the name of \code{FUN} is used
 #' @return object of class \code{stars} with accordingly reduced number of dimensions; in case \code{FUN} returns more than one value, a new dimension is created carrying the name of the function used; see the examples.
 #' @examples
 #' tif = system.file("tif/L7_ETMs.tif", package = "stars")
@@ -98,8 +103,10 @@ st_apply = function(X, MARGIN, FUN, ...) UseMethod("st_apply")
 #'   pboptions(type = "timer")
 #' }
 #' @export
-st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE, FUTURE = FALSE, rename = TRUE) {
-	fname <- paste(deparse(substitute(FUN), 50), collapse = "\n")
+st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE, FUTURE = FALSE, 
+		rename = TRUE, .fname) {
+	if (missing(.fname))
+		.fname <- paste(deparse(substitute(FUN), 50), collapse = "\n")
 	if (is.character(MARGIN))
 		MARGIN = match(MARGIN, names(dim(X)))
 	dX = dim(X)[MARGIN]
@@ -131,8 +138,8 @@ st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE,
 	ret = lapply(X, fn, ...) 
 	dim_ret = dim(ret[[1]])
 	ret = if (length(dim_ret) == length(MARGIN)) { # FUN returned a single value
-			if (length(ret) == 1 && rename && make.names(fname) == fname)
-				ret = setNames(ret, fname)
+			if (length(ret) == 1 && rename && make.names(.fname) == .fname)
+				ret = setNames(ret, .fname)
 			st_stars(ret, st_dimensions(X)[MARGIN])
 		} else { # FUN returned multiple values: need to set dimension name & values
 			dim_no_margin = dim(X)[-MARGIN]
@@ -147,7 +154,7 @@ st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE,
 			} else {
 				orig = st_dimensions(X)[MARGIN]
 				r = attr(orig, "raster")
-				dims = c(structure(list(list()), names = fname), orig)
+				dims = c(structure(list(list()), names = .fname), orig)
 				dims[[1]] = if (!is.null(dimnames(ret[[1]])[[1]])) # FUN returned named vector:
 						create_dimension(values = dimnames(ret[[1]])[[1]])
 					else