+++ R/tidyverse.R
@@ -42,7 +42,7 @@ filter.stars <- function(.data, ...) {
 
 #' @name dplyr
 filter.stars_proxy = function(.data, ...) {
-	collect(.data, match.call(), "filter", ".data")
+	collect(.data, match.call(), "filter", ".data", env = environment())
 }
 
 
@@ -54,7 +54,18 @@ mutate.stars <- function(.data, ...) {
 
 #' @name dplyr
 mutate.stars_proxy = function(.data, ...) {
-	collect(.data, match.call(), "mutate", ".data")
+	collect(.data, match.call(), "mutate", ".data", env = environment())
+}
+
+#' @name dplyr
+transmute.stars <- function(.data, ...) {
+	ret = dplyr::transmute(to_df(.data), ...)
+	st_as_stars(set_dim(ret, dim(.data)), dimensions = st_dimensions(.data))
+}
+
+#' @name dplyr
+transmute.stars_proxy = function(.data, ...) {
+	collect(.data, match.call(), "transmute", ".data", env = environment())
 }
 
 
@@ -69,7 +80,7 @@ select.stars <- function(.data, ...) {
 
 #' @name dplyr
 select.stars_proxy = function(.data, ...) {
-	collect(.data, match.call(), "select", ".data")
+	collect(.data, match.call(), "select", ".data", env = environment())
 }
 
 #' @param var see \link[dplyr]{pull}
@@ -85,7 +96,7 @@ pull.stars = function (.data, var = -1) {
 
 #' @name dplyr
 pull.stars_proxy = function(.data, ...) {
-	collect(.data, match.call(), "pull", ".data")
+	collect(.data, match.call(), "pull", ".data", env = environment())
 }
 
 #' @name dplyr
@@ -131,7 +142,7 @@ slice.stars <- function(.data, along, index, ..., drop = length(index) == 1) {
 
 #' @name dplyr
 slice.stars_proxy = function(.data, ...) {
-	collect(.data, match.call(), "slice", ".data")
+	collect(.data, match.call(), "slice", ".data", env = environment())
 }
 
 #' @name st_coordinates
@@ -238,6 +249,8 @@ register_all_s3_methods = function() {
 	register_s3_method("dplyr", "pull", "stars_proxy")
 	register_s3_method("dplyr", "slice", "stars")
 	register_s3_method("dplyr", "slice", "stars_proxy")
+	register_s3_method("dplyr", "transmute", "stars")
+	register_s3_method("dplyr", "transmute", "stars_proxy")
 	register_s3_method("lwgeom", "st_transform_proj", "stars")
 	register_s3_method("sf", "st_join", "stars")
 	register_s3_method("spatstat", "as.owin", "stars")