+++ R/write.R
@@ -19,7 +19,7 @@ reset_sub = function(x) {
 }
 
 st_write.stars = function(obj, dsn, layer, ...) {
-	.Deprecated("read_stars") # nocov
+	.Deprecated("write_stars") # nocov
 }
 
 
@@ -67,14 +67,19 @@ write_stars.stars_proxy = function(obj, dsn, layer = 1, ..., driver = detect.dri
 		options = character(0), type = "Float32", NA_value = NA_real_, 
 		chunk_size = c(dim(obj)[1], floor(25e6 / dim(obj)[1])), progress = TRUE) {
 
-	if (missing(layer) && length(obj) > 1)
-		warning("all but first attribute are ignored")
-	if (layer != 1)
-		stop("only first attribute of a stars_proxy object can be written; consider using merge")
+	if (missing(layer)) {
+		if (length(obj) > 1)
+			warning("all but first attribute are ignored")
+	} else {
+		warning("specifying layer may not work in write_stars()")
+		obj = obj[layer]
+	}
 	if (length(obj[[1]]) > 1) { # collapse bands:
 		out_file = tempfile(fileext = ".vrt")
-		gdal_utils("buildvrt", x[[1]], out_file, options = "-separate")
-		x[[1]] = out_file
+		gdal_utils("buildvrt", obj[[1]], out_file, options = "-separate")
+		obj = unclass(obj)
+		obj[[1]] = out_file
+		obj = structure(obj, class = c("stars_proxy", "stars"))
 	}
 	if (progress) {
 		pb = txtProgressBar()
@@ -82,50 +87,47 @@ write_stars.stars_proxy = function(obj, dsn, layer = 1, ..., driver = detect.dri
 	}
 
 	dim_obj = dim(obj)
-	if (prod(chunk_size) > prod(dim_obj[1:2])) {
+	if (prod(chunk_size) > prod(dim_obj[1:2]))
 		write_stars(st_as_stars(obj), dsn, layer, ..., driver = driver, options = options,
 			type = type, NA_value = NA_value)
-		return(invisible(obj))
-	}
+	else { # write chunked:
+		di = st_dimensions(obj)
+		if (di[[1]]$from > 1 || di[[2]]$from > 1)
+			message("chunked writing may not work for subsetted rasters: in case of failure use write_stars(st_as_stars(object))")
 
-	# write chunks:
-	di = st_dimensions(obj)
-	if (di[[1]]$from > 1 || di[[2]]$from > 1)
-		message("chunked writing may not work for subsetted rasters: in case of failure use write_stars(st_as_stars(object))")
+		created = FALSE
 
-	created = FALSE
-
-	ncol = ceiling(dim_obj[1] / chunk_size[1])
-	nrow = ceiling(dim_obj[2] / chunk_size[2])
-	for (col in 1:ncol) { 
-		di[[1]]$from = 1 + (col - 1) * chunk_size[1]
-		di[[1]]$to   = min(col * chunk_size[1], dim_obj[1])
-		for (row in 1:nrow) {
-			di[[2]]$from = 1 + (row - 1) * chunk_size[2]
-			di[[2]]$to   = min(row * chunk_size[2], dim_obj[2])
-			chunk = st_as_stars(structure(obj, dimensions = di))
-			if (! created) { # create:
-				d = st_dimensions(chunk)
-				d_obj = st_dimensions(obj)
-				d[[1]]$from = d[[2]]$from = 1
-				d[[1]]$to = d_obj[[1]]$to
-				d[[2]]$to = d_obj[[2]]$to
-				# reset dimensions 1/2 to original:
-				sf::gdal_write(structure(obj, dimensions = d), ..., file = dsn, driver = driver, options = options, 
-					type = type, NA_value = NA_value, geotransform = get_geotransform(obj)) # branches on stars_proxy
+		ncol = ceiling(dim_obj[1] / chunk_size[1])
+		nrow = ceiling(dim_obj[2] / chunk_size[2])
+		for (col in 1:ncol) {
+			di[[1]]$from = 1 + (col - 1) * chunk_size[1]
+			di[[1]]$to   = min(col * chunk_size[1], dim_obj[1])
+			for (row in 1:nrow) {
+				di[[2]]$from = 1 + (row - 1) * chunk_size[2]
+				di[[2]]$to   = min(row * chunk_size[2], dim_obj[2])
+				chunk = st_as_stars(structure(obj, dimensions = di))
+				if (! created) { # create:
+					d = st_dimensions(chunk)
+					d_obj = st_dimensions(obj)
+					d[[1]]$from = d[[2]]$from = 1
+					d[[1]]$to = d_obj[[1]]$to
+					d[[2]]$to = d_obj[[2]]$to
+					# reset dimensions 1/2 to original:
+					sf::gdal_write(structure(obj, dimensions = d), ..., file = dsn, driver = driver, options = options, 
+						type = type, NA_value = NA_value, geotransform = get_geotransform(obj)) # branches on stars_proxy
+				}
+				created = TRUE
+				write_stars(chunk, dsn = dsn, layer = layer, driver = driver,
+					options = options, type = type, update = TRUE, NA_value = NA_value)
+				if (progress)
+					setTxtProgressBar(pb, ((col-1) * nrow + row) / (ncol * nrow))
 			}
-			created = TRUE
-			write_stars(chunk, dsn = dsn, layer = layer, driver = driver,
-				options = options, type = type, update = TRUE)
-			if (progress)
-				setTxtProgressBar(pb, ((col-1) * nrow + row) / (ncol * nrow))
 		}
 	}
 	if (progress)
 		close(pb)
 	invisible(obj)
-}
-
+} 
 #' @name write_stars
 #' @export
 #' @param filename character; used for guessing driver short name based on file 
@@ -137,34 +139,41 @@ detect.driver = function(filename) { #nocov start
 	ext <- tolower(tools::file_ext(filename))
 	if (nchar(ext) < 2) {
 		warning("file without extension, using driver GTiff", call. = FALSE)
-		"GTiff"
+		'GTiff'
 	} else {
-		if (ext == 'tif' | ext == 'tiff') { return('GTiff')
-		} else if (ext == 'grd') { return('raster')
-		} else if (ext == 'asc') { return('ascii')
-		} else if (ext == 'nc' || ext == 'cdf' || ext == 'ncdf') { return('netcdf')
-		} else if (ext == 'kml') { return('KML')
-		} else if (ext == 'kmz') { return('KML')		
-		} else if (ext == 'big') { return('big.matrix')
-		} else if (ext == 'sgrd') { return('SAGA')
-		} else if (ext == 'sdat') { return('SAGA')
-		} else if (ext == 'bil') { return('BIL')
-		} else if (ext == 'bsq') { return('BSQ')
-		} else if (ext == 'bip') { return('BIP')
-		} else if (ext == 'bmp') { return('BMP') 
-		} else if (ext == 'gen') { return('ADRG') 
-		} else if (ext == 'bt') { return('BT') 
-		} else if (ext == 'envi') { return('ENVI')
-		} else if (ext == 'ers') { return('ERS') 
-		} else if (ext == 'img') { return( 'HFA') 
-		} else if (ext == 'rst') { return('RST') 
-		} else if (ext == 'mpr') { return('ILWIS')
-		} else if (ext == 'rsw') { return('RMF')
-		} else if (ext == 'flt') { return('EHdr')
-		} else { 
-			warning('extension ', ext, ' is unknown. Using default driver GTiff.')
-			return('GTiff') 
+		switch(ext, 
+			tif  = ,
+			tiff = 'GTiff',
+			grd  = 'rraster',
+			asc  = 'ascii',
+			nc   = ,
+			cdf  = ,
+			ncdf = 'netcdf',
+			kml  = ,
+			kmz  = 'KML',
+			big  = 'big.matrix',
+			sgrd = ,
+			sdat = 'SAGA',
+			bil  = 'BIL',
+			bsq  = 'BSQ',
+			bip  = 'BIP',
+			bmp  = 'BMP',
+			gen  = 'ADRG',
+			bt   = 'BT',
+			envi = 'ENVI',
+			ers  = 'ERS',
+			img  = 'HFA',
+			rst  = 'RST',
+			mpr  = 'ILWIS',
+			rsw  = 'RMF',
+			flt  = 'EHdr',
+			gpkg = 'GPKG',
+			{
+				warning("extension ", ext, " is unknown. Using default driver GTiff.")
+				'GTiff'
+			}
+		)
 		}
-	}
 }
 #nocov end
+